final-commit

+++ index.php
@@ -12,32 +12,54 @@
     $alert_result = false;
     $message = '';
     if(isset($_POST)) {
-        foreach($_POST as $key => $value){
-            if($key[0] == 'D') {
-                $id = substr($key, 1);
-                $query = "delete from tasks where id=$id";
+        if (isset($_POST['add_new_task'])) {
+            $name = $_POST['name_of_task'];
+            if (strlen($name) > 0) {
+                $query = "insert into `tasks` (`id`, `task`, `status`) values (NULL, '$name', '0');";
                 $alert_result = $conn->query($query);
                 if($alert_result) {
-                    $message = 'Deleted successfully';
+                        $message = 'Added successfully';
                 }
             }
-            else if ($key[0] == 'C') {
-                $id = substr($key, 1);
-                $query = "update tasks set status = 1 where id=$id";
-                $alert_result = $conn->query($query);
-                if($alert_result) {
-                    $message = 'Marked as complete';
-                }   
-            }
-            else if ($key[0] == 'U') {
-                $id = substr($key, 1);
-                $newname = $_POST['newname'];
-
-                $query = "update tasks set task=\"$newname\" where id=$id";
-                $alert_result = $conn->query($query);
-                if($alert_result) {
-                    $message = 'Updated task';
-                } 
+        }
+        else {
+            foreach($_POST as $key => $value) {
+                if($key[0] == 'D') {
+                    $id = substr($key, 1);
+                    $query = "delete from tasks where id=$id";
+                    $alert_result = $conn->query($query);
+                    if($alert_result) {
+                        $message = 'Deleted successfully';
+                    }
+                }
+                else if ($key[0] == 'C') {
+                    $id = substr($key, 1);
+                    $query = "update tasks set status = 1 where id=$id";
+                    $alert_result = $conn->query($query);
+                    if($alert_result) {
+                        $message = 'Marked as complete';
+                    }   
+                }
+                else if ($key[0] == 'U') {
+                    $id = substr($key, 1);
+                    $newname = $_POST['newname'];
+    
+                    $query = "update tasks set task=\"$newname\" where id=$id";
+                    $alert_result = $conn->query($query);
+                    if($alert_result) {
+                        $message = 'Updated task';
+                    } 
+                }
+                else if ($key[0] == 'I') {
+                    $id = substr($key, 1);
+    
+                    $query = "update tasks set status=0 where id=$id";
+                    $alert_result = $conn->query($query);
+                    if($alert_result) {
+                        $message = 'Marked as incomplete';
+                    } 
+                }
+        
             }
         }
     }
@@ -49,6 +71,13 @@
     <head>
         <title>Workshop!</title>
         <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.4/css/bootstrap.min.css">
+        <style>
+            body {
+                width: 80%;
+                margin-left: 10%;
+                margin-top: 5%;
+            }
+        </style>
     </head>
     <body>
         <form action="index.php" method="POST">
@@ -58,8 +87,11 @@
             }
             else {
                 if ($alert_result) {
-                    echo "</div>
-                        <div class=\"alert alert-success alert-dismissible\" role=\"alert\">".$message."</div>";
+                    echo "<div class=\"alert alert-success alert-dismissible\" role=\"alert\">".$message;
+                    echo "<button type=\"button\" class=\"close\" data-dismiss=\"success\" aria-label=\"Close\">
+                            <span aria-hidden=\"true\">&times;</span>
+                         </button>";
+                    echo "</div>";
                 }
                 echo "
                     <table class=\"table table-striped table-bordered table-hover\">
@@ -87,19 +119,24 @@
                     echo "<center><td>";
                     if ($row['status'] == '0') {
                         echo "<button type=\"submit\" class=\"btn btn-success\" name=\"C".$row['id']."\">
-                                    <span class=\"  glyphicon glyphicon-ok\"></span>
+                                    <span class=\"glyphicon glyphicon-ok\"></span>
                                 </button>";
                     }
-                        echo "
-                                <button type=\"submit\" class=\"btn btn-primary\" name=\"E".$row['id']."\">
-                                    <span class=\"  glyphicon glyphicon-edit\"></span>
-                                </button>
-                                <button type=\"submit\" class=\"btn btn-danger\" name=\"D".$row['id']."\">
-                                    <span class=\"glyphicon glyphicon-remove\"></span>
-                                </button>
-                            </td>
-                        </center>
-                        ";
+                    else {
+                        echo "<button type=\"submit\" class=\"btn btn-success\" name=\"I".$row['id']."\">
+                                    <span class=\"glyphicon glyphicon-refresh\"></span>
+                                </button>";
+                    }
+                    echo "
+                            <button type=\"submit\" class=\"btn btn-primary\" name=\"E".$row['id']."\">
+                                <span class=\"glyphicon glyphicon-edit\"></span>
+                            </button>
+                            <button type=\"submit\" class=\"btn btn-danger\" name=\"D".$row['id']."\">
+                                <span class=\"glyphicon glyphicon-remove\"></span>
+                            </button>
+                        </td>
+                    </center>
+                    ";
                     echo "</tr>";
                 }
                 echo "
@@ -115,17 +152,19 @@
                         $result = $conn->query($query);
                         $row = $result->fetch_assoc();
 
-
-                        echo "<form action=\"index.php\" method=\"POST\">";
                         echo "<strong> Task Name: </strong>";
                         echo "<input type=\"text\" name=\"newname\" value=\"".$row['task']."\"></input>";
                         echo "<button type=\"submit\" name=\"U".$id."\">Update</button>";
-                        echo "</form>";
                     }
                 }
             }
-            unset($_POST);
             ?>
+            <div>
+                <h3>New task</h3>
+                <p>Name: </p>
+                <input type="text" name="name_of_task">
+                <button type="submit" name="add_new_task">Add</button>  
+            </div>
         </form>
         <script src="bootstrap.min.js"></script>
     </body>
